Step 8: Introduce SymbolParser port and ClangSymbolParser adapter
