+++ clang_index_mcp/_symbols/ports/__init__.py
@@ -1,5 +1,6 @@
 """Ports (interfaces) for the symbols/domain layer."""
 
 from .call_graph import CallGraphPort
+from .parser import ParseResult, SymbolParser
 
-__all__ = ["CallGraphPort"]
+__all__ = ["CallGraphPort", "ParseResult", "SymbolParser"]
